// PlanetPulseFrontend/src/app/models/transport-mode.ts
export enum TransportMode {
    DomesticFlight = 'Domestic_flight',
    DieselCar = 'Diesel_car',
    PetrolCar = 'Petrol_car',
    ShortHaulFlight = 'Short_haul_flight',
    LongHaulFlight = 'Long_haul_flight',
    Motorbike = 'Motorbike',
    Bus = 'Bus',
    BusCity = 'Bus_city',
    PlugInHybrid = 'Plug_in_hybrid',
    ElectricCar = 'Electric_car',
    NationalRail = 'National_rail',
    Tram = 'Tram',
    Underground = 'Underground',
    FerryFootPassenger = 'Ferry_foot_passenger',
    EBike = 'e_bike',
}
  